Move POJOs into RoutePlannerRestHandler
