+++ LavalinkServer/src/main/java/lavalink/server/routeplanner/RoutePlannerRestHandler.kt
@@ -4,7 +4,6 @@ import com.sedmelluq.lava.extensions.youtuberotator.planner.AbstractRoutePlanner
 import com.sedmelluq.lava.extensions.youtuberotator.planner.NanoIpRoutePlanner
 import com.sedmelluq.lava.extensions.youtuberotator.planner.RotatingIpRoutePlanner
 import com.sedmelluq.lava.extensions.youtuberotator.planner.RotatingNanoIpRoutePlanner
-import lavalink.server.routeplanner.pojo.*
 import org.json.JSONObject
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -107,4 +106,36 @@ class RoutePlannerRestHandler(val routePlanner: AbstractRoutePlanner?) {
             }
         }
     }
+
+    data class RoutePlannerStatus(val `class`: String, val details: AbstractRoutePlannerStatus)
+
+    interface AbstractRoutePlannerStatus
+    data class GenericRoutePlannerStatus(
+            val ipBlock: IpBlockStatus,
+            val failingAddresses: List<FailingAddress>
+    ) : AbstractRoutePlannerStatus
+
+    data class RotatingIpRoutePlannerStatus(
+            val ipBlock: IpBlockStatus,
+            val failingAddresses: List<FailingAddress>,
+            val rotateIndex: String,
+            val ipIndex: String,
+            val currentAddress: String
+    ) : AbstractRoutePlannerStatus
+
+    data class NanoIpRoutePlannerStatus(
+            val ipBlock: IpBlockStatus,
+            val failingAddresses: List<FailingAddress>,
+            val currentAddressIndex: String
+    ) : AbstractRoutePlannerStatus
+
+    data class RotatingNanoIpRoutePlannerStatus(
+            val ipBlock: IpBlockStatus,
+            val failingAddresses: List<FailingAddress>,
+            val blockIndex: String,
+            val currentAddressIndex: String
+    ) : AbstractRoutePlannerStatus
+
+    data class FailingAddress(val failingAddress: String, val failingTimestamp: Long, val failingTime: String)
+    data class IpBlockStatus(val type: String, val size: String)
 }
