--- LavalinkServer/src/main/java/lavalink/server/routeplanner/pojo/FailingAddress.kt
@@ -1,5 +1,0 @@
-package lavalink.server.routeplanner.pojo
-
-import java.net.InetAddress
-
-data class FailingAddress(val failingAddress: String, val failingTimestamp: Long, val failingTime: String)
