--- LavalinkServer/src/main/java/lavalink/server/routeplanner/pojo/IpBlockStatus.kt
@@ -1,3 +1,0 @@
-package lavalink.server.routeplanner.pojo
-
-data class IpBlockStatus(val type: String, val size: String)
