--- LavalinkServer/src/main/java/lavalink/server/routeplanner/pojo/RoutePlannerStatus.kt
@@ -1,3 +1,0 @@
-package lavalink.server.routeplanner.pojo
-
-data class RoutePlannerStatus(val `class`: String, val details: AbstractRoutePlannerStatus)
